+++ p2panda-rs/src/schema/system/schema_field_definition.rs
@@ -1,23 +1,23 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-use std::collections::BTreeMap;
-
 use once_cell::sync::Lazy;
 
 use crate::schema::error::SchemaIdError;
-use crate::schema::{FieldType, Schema, SchemaId};
+use crate::schema::{FieldType, Schema, SchemaDescription, SchemaFields, SchemaId};
 
 const DESCRIPTION: &str = "Define fields for application data schemas.";
 
 pub static SCHEMA_FIELD_DEFINITION_V1: Lazy<Schema> = Lazy::new(|| {
-    let mut fields = BTreeMap::new();
+    let fields = SchemaFields::new(&[("name", FieldType::String), ("type", FieldType::String)])
+        // Unwrap as we know the fields are valid.
+        .unwrap();
 
-    fields.insert("name".to_string(), FieldType::String);
-    fields.insert("type".to_string(), FieldType::String);
+    // We can unwrap here as we know the schema definition is valid.
+    let description = SchemaDescription::new(DESCRIPTION).unwrap();
 
     Schema {
         id: SchemaId::SchemaFieldDefinition(1),
-        description: DESCRIPTION.to_owned(),
+        description,
         fields,
     }
 });