+++ p2panda-rs/src/schema/validate/schema_definition.rs
@@ -25,7 +25,7 @@ pub fn validate_name(value: &str) -> bool {
 ///
 /// 1. It consists of unicode characters
 /// 2. ... and must be at most 256 characters long
-fn validate_description(value: &str) -> bool {
+pub fn validate_description(value: &str) -> bool {
     value.chars().count() <= 256
 }