+++ p2panda-rs/src/test_utils/constants.rs
@@ -4,7 +4,7 @@
 use crate::operation::{
     OperationValue, PinnedRelation, PinnedRelationList, Relation, RelationList,
 };
-use crate::schema::Schema;
+use crate::schema::{FieldType, Schema};
 use crate::test_utils::fixtures::{document_id, document_view_id, schema_fields, schema_id};
 
 /// Hash value, used when a hash is needed for testing. It's the default hash in fixtures
@@ -73,14 +73,20 @@ pub fn test_fields() -> Vec<(&'static str, OperationValue)> {
 /// Derived from the test fields defined above.
 pub fn schema() -> Schema {
     let id = schema_id(SCHEMA_ID);
-    Schema::new(&id, "Test schema", schema_fields(test_fields(), id.clone())).unwrap()
+    let fields = schema_fields(test_fields(), id.clone());
+    let fields: Vec<(&str, FieldType)> = fields
+        .iter()
+        .map(|(name, field_type)| (name.as_str(), field_type.to_owned()))
+        .collect();
+
+    Schema::new(&id, "Test schema", &fields).unwrap()
 }
 
 #[cfg(test)]
 mod tests {
     use crate::hash::Hash;
     use crate::operation::OperationId;
-    use crate::schema::SchemaId;
+    use crate::schema::{SchemaId, SchemaName};
 
     use super::*;
 
@@ -93,7 +99,8 @@ mod tests {
     #[test]
     fn default_schema() {
         let venue_schema_hash: OperationId = Hash::new_from_bytes(&[3, 2, 1]).into();
-        let schema = SchemaId::new_application("venue", &venue_schema_hash.into());
+        let venue_schema_name = SchemaName::new("venue").expect("Valid schema name");
+        let schema = SchemaId::new_application(&venue_schema_name, &venue_schema_hash.into());
         assert_eq!(schema.to_string(), SCHEMA_ID)
     }
 }